commonfmt: fix some clippy warnings

+++ nihav-commonfmt/src/codecs/pcm.rs
@@ -25,10 +25,10 @@ fn get_default_chmap(nch: u8) -> NAChannelMap {
 fn get_duration(ainfo: &NAAudioInfo, duration: Option<u64>, data_size: usize) -> u64 {
     if duration == None {
         let size_bits = (data_size as u64) * 8;
-        let blk_size = (ainfo.get_channels() as u64) * (ainfo.get_format().get_bits() as u64);
+        let blk_size = u64::from(ainfo.get_channels()) * u64::from(ainfo.get_format().get_bits());
         size_bits / blk_size
     } else {
-        duration.unwrap() as u64
+        duration.unwrap()
     }
 }
